// src/types/custom-d-t.ts
export interface AppContextType {
  openOffcanvas: boolean;
  setOpenOffcanvas: React.Dispatch<React.SetStateAction<boolean>>;
  openCartOffcanvas: boolean;
  setOpenCartOffcanvas: React.Dispatch<React.SetStateAction<boolean>>;
  toggleCartOffcanvas: () => void;
  toggleOffcanvas: () => void;
  toggleSearch: () => void;
  toggleModal: () => void;
  openModal: boolean;
  openSearch: boolean;
  setOpenSearch: React.Dispatch<React.SetStateAction<boolean>>;
}